pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vits"
version = "0.1.0"
description = "Vision-Informed Time Series anomaly detection (NeurIPS 2026)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Anonymous" }]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "torch>=2.1",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
vision = [
    "transformers>=4.40",
    "timm>=0.9",
    "Pillow>=10.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.5",
]
full = [
    "vits[vision,dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
